match: ScanGroup's byte loop is unsigned; DrawTextLines copies whole RECTs
CTriggerMgr::ScanGroup 0x7a760 (89.01 -> 89.38): the m_byteArr serialisation loop
guards are `jbe`/`jb` in retail and `jle`/`jl` in ours - the count is unsigned.
Residue is exactly 14 instructions: retail duplicates the epilogue at both
NULL-guard early-outs where cl folds ours onto one shared tail.

CFontConfig::DrawTextLines 0x22360 (85.40 -> 89.51):
* `cur` and `work` are whole-struct copies (`RECT cur = *rect; RECT work = *rect;`).
  MSVC expands each as four load/store pairs and does NOT CSE across the two
  expansions - retail reloads every field for the second copy through a second base
  register, which field-by-field assignment cannot reproduce (cl CSEs those).
* The clamp is `(count < m_list.GetCount()) ? count : m_list.GetCount()`, so the
  taken branch is the `jge` to the GetCount() arm, not our inverted spelling.

+++ src/Gruntz/TriggerMgr.cpp
@@ -1258,9 +1258,9 @@ i32 CTriggerMgr::ScanGroup(CFileMemBase* ar) {
     ar->Write(m_cellFlag, 0xf0);
     ar->Write(m_gruntzExitedByPlayer, 0x10);
     ar->Write(m_gruntzLostByPlayer, 0x10);
-    i32 cnt = m_byteArr.GetSize();
+    u32 cnt = static_cast<u32>(m_byteArr.GetSize());
     ar->Write(&cnt, sizeof(cnt));
-    for (i32 i = 0; i < cnt; i++) {
+    for (u32 i = 0; i < cnt; i++) {
         u8 b = m_byteArr.GetData()[i];
         ar->Write(&b, sizeof(b));
     }
